# openapi.json
{"openapi":"3.1.0","info":{"title":"tAPI - Todo API","description":"\n    A RESTful API for managing todos with features including:\n    \n    * **Create** todos with title, description, due dates, and tags\n    * **List** todos with filtering and search capabilities\n    * **Update** todo items\n    * **Delete** todos\n    \n    All endpoints require API key authentication via the `X-API-Key` header.\n    ","contact":{"name":"tAPI Support"},"license":{"name":"MIT"},"version":"1.0.0"},"paths":{"/todos/":{"post":{"tags":["todos"],"summary":"Create a new todo","description":"Create a new todo item with title, description, due date, priority, and tags.","operationId":"create_todo_todos__post","security":[{"APIKeyHeader":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TodoCreate"}}}},"responses":{"201":{"description":"The created todo item","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TodoRead"}}}},"401":{"description":"Unauthorized - Missing API key header"},"403":{"description":"Forbidden - Invalid API key"},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["todos"],"summary":"List todos","description":"Retrieve a list of todos with optional filtering and search capabilities.","operationId":"list_todos_todos__get","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"q","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Search query to filter todos by title (case-insensitive partial match)","title":"Q"},"description":"Search query to filter todos by title (case-insensitive partial match)","example":"project"},{"name":"tag","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter todos by a specific tag","title":"Tag"},"description":"Filter todos by a specific tag","example":"work"},{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter todos by status (e.g., 'todo', 'in_progress', 'done')","title":"Status"},"description":"Filter todos by status (e.g., 'todo', 'in_progress', 'done')","example":"todo"},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":1000,"description":"Maximum number of todos to return (max: 1000)","default":100,"title":"Limit"},"description":"Maximum number of todos to return (max: 1000)","example":50}],"responses":{"200":{"description":"List of todo items matching the criteria","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TodoRead"},"title":"Response List Todos Todos  Get"}}}},"401":{"description":"Unauthorized - Missing API key header"},"403":{"description":"Forbidden - Invalid API key"},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/todos/{item_id}":{"get":{"tags":["todos"],"summary":"Get a todo by ID","description":"Retrieve a specific todo item by its unique identifier.","operationId":"get_todo_todos__item_id__get","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"item_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Item Id"}}],"responses":{"200":{"description":"The todo item","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TodoRead"}}}},"401":{"description":"Unauthorized - Missing API key header"},"403":{"description":"Forbidden - Invalid API key"},"404":{"description":"Todo not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["todos"],"summary":"Update a todo","description":"Partially update a todo item. Only provided fields will be updated.","operationId":"update_todo_todos__item_id__patch","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"item_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Item Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TodoUpdate"}}}},"responses":{"200":{"description":"The updated todo item","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TodoRead"}}}},"401":{"description":"Unauthorized - Missing API key header"},"403":{"description":"Forbidden - Invalid API key"},"404":{"description":"Todo not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["todos"],"summary":"Delete a todo","description":"Delete a todo item by its unique identifier.","operationId":"delete_todo_todos__item_id__delete","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"item_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Item Id"}}],"responses":{"204":{"description":"No content on successful deletion"},"401":{"description":"Unauthorized - Missing API key header"},"403":{"description":"Forbidden - Invalid API key"},"404":{"description":"Todo not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"tags":["health"],"summary":"Health check endpoint","description":"Returns the health status of the API","operationId":"health_health_get","responses":{"200":{"description":"API health status","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"TodoCreate":{"properties":{"title":{"type":"string","title":"Title","description":"The title of the todo item","example":"Complete project documentation"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description","description":"Detailed description of the todo","example":"Write comprehensive API documentation with examples"},"due_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Due At","description":"Due date and time for the todo","example":"2024-12-31T23:59:59"},"estimated_minutes":{"anyOf":[{"type":"integer","minimum":0.0},{"type":"null"}],"title":"Estimated Minutes","description":"Estimated time to complete in minutes","example":120},"priority":{"anyOf":[{"type":"integer","maximum":5.0,"minimum":1.0},{"type":"null"}],"title":"Priority","description":"Priority level (1=highest, 5=lowest)","default":3,"example":2},"tags":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Tags","description":"List of tags for categorizing the todo","example":["work","urgent","documentation"]}},"type":"object","required":["title"],"title":"TodoCreate","description":"Schema for creating a new todo item.","example":{"description":"Write comprehensive API documentation with examples","due_at":"2024-12-31T23:59:59","estimated_minutes":120,"priority":2,"tags":["work","urgent","documentation"],"title":"Complete project documentation"}},"TodoRead":{"properties":{"title":{"type":"string","title":"Title","description":"The title of the todo item","example":"Complete project documentation"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description","description":"Detailed description of the todo","example":"Write comprehensive API documentation with examples"},"due_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Due At","description":"Due date and time for the todo","example":"2024-12-31T23:59:59"},"estimated_minutes":{"anyOf":[{"type":"integer","minimum":0.0},{"type":"null"}],"title":"Estimated Minutes","description":"Estimated time to complete in minutes","example":120},"priority":{"anyOf":[{"type":"integer","maximum":5.0,"minimum":1.0},{"type":"null"}],"title":"Priority","description":"Priority level (1=highest, 5=lowest)","default":3,"example":2},"tags":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Tags","description":"List of tags for categorizing the todo","example":["work","urgent","documentation"]},"id":{"type":"string","format":"uuid","title":"Id","description":"Unique identifier for the todo item","example":"123e4567-e89b-12d3-a456-426614174000"},"created_at":{"type":"string","format":"date-time","title":"Created At","description":"Timestamp when the todo was created","example":"2024-01-15T10:30:00"},"updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Updated At","description":"Timestamp when the todo was last updated","example":"2024-01-16T14:20:00"}},"type":"object","required":["title","id","created_at"],"title":"TodoRead","description":"Schema for reading a todo item with all fields including metadata.","example":{"created_at":"2024-01-15T10:30:00","description":"Write comprehensive API documentation with examples","due_at":"2024-12-31T23:59:59","estimated_minutes":120,"id":"123e4567-e89b-12d3-a456-426614174000","priority":2,"status":"todo","tags":["work","urgent","documentation"],"title":"Complete project documentation","updated_at":"2024-01-16T14:20:00"}},"TodoUpdate":{"properties":{"title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title","description":"The title of the todo item","example":"Updated project documentation"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description","description":"Detailed description of the todo","example":"Updated description"},"due_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Due At","description":"Due date and time for the todo","example":"2024-12-31T23:59:59"},"estimated_minutes":{"anyOf":[{"type":"integer","minimum":0.0},{"type":"null"}],"title":"Estimated Minutes","description":"Estimated time to complete in minutes","example":90},"status":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status","description":"Status of the todo (e.g., 'todo', 'in_progress', 'done')","example":"in_progress"},"priority":{"anyOf":[{"type":"integer","maximum":5.0,"minimum":1.0},{"type":"null"}],"title":"Priority","description":"Priority level (1=highest, 5=lowest)","example":1},"tags":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Tags","description":"List of tags for categorizing the todo","example":["work","completed"]}},"type":"object","title":"TodoUpdate","description":"Schema for updating a todo item. All fields are optional.","example":{"priority":1,"status":"in_progress","tags":["work","completed"],"title":"Updated project documentation"}},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"APIKeyHeader":{"type":"apiKey","description":"API key for authentication. Get your API key from the administrator.","in":"header","name":"X-API-Key"}}}}